requirements.txt
amqp==5.0.9
argcomplete==1.12.3
arrow==1.2.2
asgiref==3.5.0
astroid==2.11.5
asttokens==2.0.5
attrs==21.2.0
Automat==20.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
Beaker==1.10.0
beautifulsoup4==4.11.0
billiard==3.6.4.0
binaryornot==0.4.4
black==22.1.0
blivet==3.4.4
blivet-gui==2.3.0
Brlapi==0.8.2
celery==5.2.6
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
chrome-gnome-shell==0.0.0
click==8.0.1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
conda==4.11.0
conda-package-handling==1.7.3
constantly==15.1.0
cookiecutter==1.7.3
cryptography==3.4.7
cssselect==1.1.0
cupshelpers==1.0
cytoolz==0.11.0
dasbus==1.6
dbus-python==1.2.18
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deluge==2.0.5
dill==0.3.5.1
distlib==0.3.2
distro==1.6.0
Django==4.0.2
django-browser-reload==1.3.0
django-cors-headers==3.11.0
django-default-imagefield==0.5.5
django-filter==21.1
django-phonenumber-field==6.3.0
django-tailwind==3.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
entrypoints==0.4
executing==0.8.2
fedora-third-party==0.8
filelock==3.3.1
flake8==4.0.1
fros==1.1
frozendict==1.2
future==0.18.2
GeoIP==1.3.2
gpg==1.15.1
gssapi==1.6.14
httpie==3.2.1
httpie-jwt-auth==0.4.0
humanize==0.5.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
ipykernel==6.9.1
ipython==8.0.1
isort==5.10.1
itemadapter==0.5.0
itemloaders==1.0.4
jedi==0.18.1
Jinja2==3.1.1
jinja2-time==0.2.0
jmespath==1.0.0
jupyter-client==7.1.2
jupyter-core==4.9.2
koji==1.29.0
kombu==5.2.3
langtable==0.0.58
lazy-object-proxy==1.7.1
libcomps==0.1.18
libtorrent==2.0.6
lxml==4.6.5
Mako==1.1.4
Markdown==3.3.6
MarkupSafe==2.0.0
matplotlib-inline==0.1.3
mccabe==0.6.1
meson==0.62.1
mod-wsgi==4.9.0
multidict==5.1.0
mypy==0.941
mypy-extensions==0.4.3
nest-asyncio==1.5.4
nftables==0.1
numpy==1.21.5
olefile==0.46
packaging==21.3
parsel==1.6.0
parso==0.8.3
Paste==3.5.0
pathspec==0.9.0
pexpect==4.8.0
phonenumberslite==8.12.50
pickleshare==0.7.5
pid==2.2.3
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.5.1
ply==3.11
poyo==0.5.0
productmd==1.33
progressbar2==3.53.2
prompt-toolkit==3.0.28
Protego==0.2.1
ptyprocess==0.6.0
pure-eval==0.2.2
pwquality==1.4.4
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.8.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.44.1
PyDispatcher==2.0.5
pyenchant==3.2.2
pyflakes==2.4.0
pygame==2.1.2
pygit2==1.6.1
Pygments==2.11.2
PyGObject==3.42.0
PyJWT==2.3.0
pykickstart==3.34
pylint==2.13.9
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.12.0
PyQt5==5.15.6
PyQt5-sip==12.9.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.1
python-dotenv==0.19.2
python-meh==0.50
python-slugify==6.1.1
python-utils==2.5.6
pytz==2021.3
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
queuelib==1.6.2
regex==2022.6.2
rencode==1.0.6
requests==2.27.1
requests-file==1.5.1
requests-ftp==0.3.1
requests-gssapi==1.2.3
requests-toolbelt==0.9.1
rich==10.7.0
rpm==4.17.0
rpmautospec==0.2.6
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
Scrapy==2.6.1
selinux==3.3
sepolicy==3.3
service-identity==21.1.0
setools==4.4.0
shodan==1.27.0
simpleaudio==1.0.4
simpleline==1.8.2
six==1.16.0
sos==4.2
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
systemd-python==234
Tempita==0.5.2
text-unidecode==1.3
tldextract==3.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.1
torbrowser-launcher==0.3.5
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
Twisted==21.7.0
typing-extensions==3.10.0.2
urllib3==1.26.7
vine==5.0.0
virtualenv==20.13.4
virtualenv-clone==0.5.7
w3lib==1.22.0
wcwidth==0.2.5
wrapt==1.14.1
XlsxWriter==3.0.3
zope.event==4.2.0
zope.interface==5.4.0
